server/my-app/requirements.txt
blinker==1.8.2
click==8.1.7
convex==0.6.0
DateTime==5.5
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.1.0
Werkzeug==3.0.4
zope.interface==7.0.3
